Revert "Move `FormulaMaterializer` to `formulaic.materializers.types` for parity with `parser`."
This reverts commit 765409e01d795b183672249b0f9118400f1b0275.
