+++ formulaic/materializers/__init__.py
@@ -1,7 +1,7 @@
 from .arrow import ArrowMaterializer
+from .base import FormulaMaterializer
 from .pandas import PandasMaterializer
 from .types import ClusterBy, FactorValues, NAAction
-from .types.formula_materializer import FormulaMaterializer
 
 __all__ = [
     "ArrowMaterializer",
