+++ formulaic/materializers/pandas.py
@@ -14,8 +14,8 @@ from formulaic.utils.cast import as_columns
 from formulaic.utils.null_handling import drop_rows as drop_nulls
 from formulaic.utils.null_handling import find_nulls
 
+from .base import FormulaMaterializer
 from .types import NAAction
-from .types.formula_materializer import FormulaMaterializer
 
 if TYPE_CHECKING:  # pragma: no cover
     from formulaic.model_spec import ModelSpec
