+++ formulaic/materializers/types/__init__.py
@@ -1,14 +1,12 @@
 from .enums import ClusterBy, NAAction
 from .evaluated_factor import EvaluatedFactor
 from .factor_values import FactorValues
-from .formula_materializer import FormulaMaterializer
 from .scoped_factor import ScopedFactor
 from .scoped_term import ScopedTerm
 
 __all__ = [
     "EvaluatedFactor",
     "FactorValues",
-    "FormulaMaterializer",
     "ClusterBy",
     "NAAction",
     "ScopedFactor",
