+++ tests/materializers/test_base.py
@@ -2,6 +2,7 @@ import pandas
 import pytest
 
 from formulaic.errors import FactorEncodingError, FormulaMaterializerNotFoundError
+from formulaic.materializers.base import FormulaMaterializer
 from formulaic.materializers.pandas import PandasMaterializer
 from formulaic.materializers.types import (
     EvaluatedFactor,
@@ -9,7 +10,6 @@ from formulaic.materializers.types import (
     ScopedFactor,
     ScopedTerm,
 )
-from formulaic.materializers.types.formula_materializer import FormulaMaterializer
 from formulaic.model_spec import ModelSpec
 from formulaic.parser.types import Factor
